replaces ParseRunner with MemoryEfficientParseRunner implementation.
